Rename old cluster_bins to cluster_bins_gmm and _loc to cluster_bins
