+++ src/hatchet/__main__.py
@@ -15,8 +15,8 @@ from hatchet.utils.genotype_snps import main as genotype_snps  # noqa: F401
 from hatchet.utils.count_alleles import main as count_alleles  # noqa: F401
 from hatchet.utils.combine_counts import main as combine_counts  # noqa: F401
 from hatchet.utils.combine_counts_fw import main as combine_counts_fw  # noqa: F401
+from hatchet.utils.cluster_bins_gmm import main as cluster_bins_gmm  # noqa: F401
 from hatchet.utils.cluster_bins import main as cluster_bins  # noqa: F401
-from hatchet.utils.cluster_bins_loc import main as cluster_bins_loc  # noqa: F401
 
 from hatchet.utils.plot_bins import main as plot_bins  # noqa: F401
 from hatchet.utils.plot_bins_1d2d import main as plot_bins_1d2d  # noqa: F401
