+++ src/hatchet/hatchet.ini
@@ -56,7 +56,7 @@ alpha = 0.1
 ss_em =
 not_compressed =
 
-[cluster_bins]
+[cluster_bins_gmm]
 outsegments =
 outbins =
 diploidbaf = 0.1
@@ -72,7 +72,7 @@ restarts = 10
 verbose = False
 disablebar = False
 
-[cluster_bins_loc]
+[cluster_bins]
 outsegments =
 outbins =
 diploidbaf = 0.1
